ref(sveltekit): Use `clack` as a CLI prompt
