+++ src/nextjs/nextjs-wizard.ts
@@ -14,10 +14,11 @@ import {
   askForSelfHosted,
   askForWizardLogin,
   confirmContinueEvenThoughNoGitRepo,
+  ensurePackageIsInstalled,
   installPackage,
   printWelcome,
   SentryProjectData,
-} from './clack-utils';
+} from '../utils/clack-utils';
 import {
   getNextjsConfigCjsAppendix,
   getNextjsConfigCjsTemplate,
@@ -27,7 +28,8 @@ import {
   getSentryConfigContents,
   getSentryExampleApiRoute,
   getSentryExamplePageContents,
-} from './templates/nextjs-templates';
+} from '../templates/nextjs-templates';
+import { getPackageDotJson } from '../utils/package-utils';
 
 interface NextjsWizardOptions {
   promoCode?: string;
@@ -44,43 +46,8 @@ export async function runNextjsWizard(
 
   await confirmContinueEvenThoughNoGitRepo();
 
-  const packageJsonFileContents = await fs.promises
-    .readFile(path.join(process.cwd(), 'package.json'), 'utf8')
-    .catch(() => {
-      clack.log.error(
-        'Could not find package.json. Make sure to run the wizard in the root of your Next.js app!',
-      );
-      abort();
-    });
-
-  let packageJson:
-    | { dependencies?: { ['@sentry/nextjs']: string; ['next']: string } }
-    | undefined = undefined;
-
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    packageJson = JSON.parse(packageJsonFileContents);
-  } catch {
-    clack.log.error(
-      'Unable to parse your package.json. Make sure it has a valid format!',
-    );
-
-    abort();
-  }
-
-  if (!packageJson?.dependencies?.['next']) {
-    const continueWithoutNext = await clack.confirm({
-      message:
-        'Next.js does not seem to be installed. Do you still want to continue?',
-      initialValue: false,
-    });
-
-    abortIfCancelled(continueWithoutNext);
-
-    if (!continueWithoutNext) {
-      abort();
-    }
-  }
+  const packageJson = await getPackageDotJson();
+  await ensurePackageIsInstalled(packageJson, 'next', 'Next.js');
 
   const { url: sentryUrl, selfHosted } = await askForSelfHosted();
